--- gencred/OWNERS
@@ -1,5 +1,0 @@
-# See the OWNERS docs at https://go.k8s.io/owners
-
-emeritus_approvers:
-- clarketm
-- fejta
